:truck: reorganise verifier tests
